Now writes to two .csv files.
All Xbox One game deals will be written to an XboxOne  csv and all Xbox 360 game deals will be written to an Xbox360 csv. This is done by simply setting a currentTable variable to 'Xbox-360'. The program will see that and run an elif statement which writes the rest of the games to the 360 csv instead.